requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.24.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
frozenlist==1.7.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
microsoft-kiota-abstractions==1.9.5
microsoft-kiota-authentication-azure==1.9.5
microsoft-kiota-http==1.9.5
microsoft-kiota-serialization-form==1.9.5
microsoft-kiota-serialization-json==1.9.5
microsoft-kiota-serialization-multipart==1.9.5
microsoft-kiota-serialization-text==1.9.5
msal==1.33.0
msal-extensions==1.3.1
msgraph-core==1.3.5
msgraph-sdk==1.40.0
multidict==6.6.4
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
propcache==0.3.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
std-uritemplate==2.0.5
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
zipp==3.23.0
